refactor(KeyframelessKit): new primitive data model for timing representation
