+++ KeyframelessKit/KeyframelessKit/KeyframelessKit.h
@@ -64,6 +64,7 @@ FOUNDATION_EXPORT const unsigned char
 #import <KeyframelessKit/KKSnapEngine.h>
 #import <KeyframelessKit/KKSquarePointOSC.h>
 #import <KeyframelessKit/KKTimingEvaluation.h>
+#import <KeyframelessKit/KKTimingLane.h>
 #import <KeyframelessKit/KKTimingStage.h>
 #import <KeyframelessKit/KKTokens.h>
 #import <KeyframelessKit/KKToolbar.h>
